Refactor: rename folder 'components' to 'ui'
